Extracting classes to separate files, adding Readme info
