+++ EpiCommonUtils/EpiCommonUtils.csproj
@@ -239,7 +239,10 @@
   <ItemGroup>
     <Compile Include="Infrastructure\DependencyInjection\StructureMapDependencyInjectionRegistrationModule.cs" />
     <Compile Include="Infrastructure\EditorDescriptors\EnumEditorDescriptor.cs" />
+    <Compile Include="Infrastructure\EditorDescriptors\EnumSelectionFactory.cs" />
     <Compile Include="Infrastructure\EditorDescriptors\EpiUnselectableAttribute.cs" />
+    <Compile Include="Infrastructure\EditorDescriptors\NullableEnumEditorDescriptor.cs" />
+    <Compile Include="Infrastructure\EditorDescriptors\NullableEnumSelectionFactory.cs" />
     <Compile Include="Infrastructure\EntryValueConverter.cs" />
     <Compile Include="Infrastructure\EnumExtensions.cs" />
     <Compile Include="Infrastructure\FeatureViewLocation\FeatureViewLocationRazorViewEngine.cs" />
