+++ EpiCommonUtils/Infrastructure/EditorDescriptors/NullableEnumEditorDescriptor.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Forte.EpiCommonUtils.Infrastructure.EditorDescriptors
+{
+    public class NullableEnumEditorDescriptor<TEnum> : EnumEditorDescriptor<TEnum, NullableEnumSelectionFactory<TEnum>> where TEnum : struct, IConvertible
+    {
+    }
+}
